requirements.txt
numpy
pandas
matplotlib
seaborn
mysql-connector
python-dotenv
scikit-learn
statsmodels
pyyaml
flask
# -e .